등록및조회 테스트

+++ beer_community/src/main/java/team_project/beer_community/dto/BeerDto.java
@@ -8,7 +8,7 @@ public class BeerDto {
     private Long id;
     private String imageUrl;
     private String beerName;
-    private float totalPoint;
+    private double totalPoint;
     private String information;
 
     public BeerDto(Beer beer) {
